+++ src/common/utils/__tests__/timeHelperFunctions.test.tsx
@@ -1,5 +1,5 @@
 import { strictEqual } from 'assert';
-import { convert12To24HourTime } from '../timeHelperFunctions';
+import { calculateStartEndTimes, convert12To24HourTime } from '../timeHelperFunctions';
 
 describe('Time Helper Functions', function () {
   describe('convert12To24HourTime', function () {
@@ -20,6 +20,18 @@ describe('Time Helper Functions', function () {
     it('should preserve the hour portion of 12:00 PM as "12"', function () {
       const noon = '12:00 PM';
       strictEqual(convert12To24HourTime(noon).split(':')[0], '12');
+    });
+  });
+  describe('calculateStartEndTimes', function () {
+    it('returns the expected start time', function () {
+      const startTime = '11:00 AM';
+      const timeslot = `${startTime}-12:00 PM`;
+      strictEqual(calculateStartEndTimes(timeslot).start, startTime);
+    });
+    it('returns the expected end time', function () {
+      const endTime = '10:30 AM';
+      const timeslot = `09:30 AM-${endTime}`;
+      strictEqual(calculateStartEndTimes(timeslot).end, endTime);
     });
   });
 });
